Ceng499 - Introduction to Machine Learning/THE1/requirements-gpu.txt
absl-py==0.6.1
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
Markdown==3.0.1
mock==2.0.0
nltk==3.3
numpy==1.15.3
pbr==5.1.0
protobuf==3.6.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
Werkzeug==0.14.1
